Replace Mirai Navigator with Mirai navigate action
