+++ packages/mirai/lib/mirai.dart
@@ -2,7 +2,6 @@ library mirai;
 
 export 'package:mirai/src/action_parsers/action_parsers.dart';
 export 'package:mirai/src/framework/framework.dart';
-export 'package:mirai/src/navigation/navigation.dart';
 export 'package:mirai/src/network/network.dart';
 export 'package:mirai/src/parsers/parsers.dart';
 export 'package:mirai/src/utils/utils.dart';
