--- packages/mirai/lib/src/navigation/navigation.dart
@@ -1,3 +1,0 @@
-library navigation;
-
-export 'mirai_navigator.dart';
